Add requirements.txt, fix sqlite build issue and add casts.

+++ LibCarla/source/carla/geom/Simplification.cpp
@@ -24,14 +24,14 @@ namespace geom {
       Simplify::Triangle t;
       t.material = 0;
       auto indices = pmesh->GetIndexes();
-      t.v[0] = (indices[i]) - 1;
-      t.v[1] = (indices[i + 1]) - 1;
-      t.v[2] = (indices[i + 2]) - 1;
+      t.v[0] = static_cast<int>(indices[i] - 1);
+      t.v[1] = static_cast<int>(indices[i + 1] - 1);
+      t.v[2] = static_cast<int>(indices[i + 2] - 1);
       Simplification.triangles.push_back(t);
     }
 
     // Reduce to the X% of the polys
-    float target_size = Simplification.triangles.size();
+    float target_size = static_cast<float>(Simplification.triangles.size());
     Simplification.simplify_mesh((target_size * simplification_percentage));
 
     pmesh->GetVertices().clear();
@@ -39,9 +39,9 @@ namespace geom {
 
     for (Simplify::Vertex& current_vertex : Simplification.vertices) {
       carla::geom::Vector3D v;
-      v.x = current_vertex.p.x;
-      v.y = current_vertex.p.y;
-      v.z = current_vertex.p.z;
+      v.x = static_cast<float>(current_vertex.p.x);
+      v.y = static_cast<float>(current_vertex.p.y);
+      v.z = static_cast<float>(current_vertex.p.z);
       pmesh->AddVertex(v);
     }
